add test on number of blocks across tables + indp table for transactions
