+++ tests/db_blocks_test.py
@@ -5,11 +5,11 @@ class CheckIntegrityOfDB(dbtest.DBintegrityTest):
     db_config_file = ".env"
 
     def test_orphan_blocks_in_block_metrics(self):
-        """ Edgy case where two blocks are assigned to the same slot (one of them is an orphan) """
+        """ Edgy case where two blocks are assigned to the same slot (one of them is an orphan) (after Merge) """
         sql_query = """
         select f_el_block_number, count(*)
         from t_block_metrics
-        where f_proposed = true
+        where f_proposed = true and f_slot > 4700012
         group by f_el_block_number
         having count(*) > 1
         """
@@ -22,19 +22,6 @@ class CheckIntegrityOfDB(dbtest.DBintegrityTest):
         select *
         from t_orphans
         where f_proposed = false
-        """
-        df = self.db.get_df_from_sql_query(sql_query)
-        self.assertNoRows(df)
-
-    def test_transactions_per_block(self):
-        """ make sure that the number of tracked transactions match the ones included in the corresponding block """
-        sql_query = """
-        select t_block_metrics.f_slot, count(distinct(f_hash))
-        from t_block_metrics
-        inner join t_transactions
-        on t_block_metrics.f_slot = t_transactions.f_slot
-        group by t_block_metrics.f_slot
-        having f_el_transactions != count(distinct(f_hash))
         """
         df = self.db.get_df_from_sql_query(sql_query)
         self.assertNoRows(df)
@@ -63,19 +50,6 @@ class CheckIntegrityOfDB(dbtest.DBintegrityTest):
         df = self.db.get_df_from_sql_query(sql_query)
         self.assertNoRows(df)
 
-    def test_missing_transactions_from_block(self):
-        """ Check if there are no blocks that is not present in the transaction table, but had transactions and was proposed """
-        sql_query = """
-        select *
-        from t_block_metrics
-        where f_slot not in (
-            select distinct(f_slot)
-            from t_transactions
-        ) and f_el_transactions > 0 and f_proposed = true
-        order by f_slot desc
-        """
-        df = self.db.get_df_from_sql_query(sql_query)
-        self.assertNoRows(df)
 
 if __name__ == '__main__':
     unittest.main()
